# specs/008-authenticated-backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Job Application Workspace Online API
  version: 0.1.0
  description: Authenticated online workspace and complete replacement operations.
servers:
  - url: https://app.example.com
security:
  - sessionCookie: []
  - desktopBearer: [workspace]
paths:
...
  /api/synchronizations/uploads:
    post:
      operationId: createUploadPreview
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: X-Sync-Direction
          required: true
          schema:
            type: string
            enum: [offline_to_online, working_copy_to_online]
        - in: header
          name: X-Last-Confirmed-Revision
          required: false
          schema: { type: integer, minimum: 0 }
      requestBody:
        required: true
        content:
          application/zip:
            schema: { type: string, contentEncoding: binary }
      responses:
        '201':
          description: Validated replacement preview; no active data changed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SynchronizationPreview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { description: Package exceeds configured limits. }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/synchronizations/{attemptId}/confirm:
    post:
      operationId: confirmOnlineReplacement
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceChecksum, expectedDestinationRevision, confirmReplacement]
              properties:
                sourceChecksum: { type: string }
                expectedDestinationRevision: { type: integer, minimum: 0 }
                confirmReplacement: { const: true }
      responses:
        '200':
          description: Completed idempotent replacement result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SynchronizationResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/synchronizations/{attemptId}:
    delete:
      operationId: cancelSynchronization
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Preview cancelled; active workspace unchanged. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    desktopBearer:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://app.example.com/api/auth/oauth2/authorize
          tokenUrl: https://app.example.com/api/auth/oauth2/token
          scopes:
            workspace: Access the owner's workspace.
...
  responses:
    Unauthorized:
      description: Missing, expired, revoked, or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource absent or not owned by the authenticated account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RevisionConflict:
      description: Expected revision is stale; reload or create a new synchronization preview.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request or workspace package failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            issues:
              type: array
              items:
                type: object
                required: [path, message]
                properties:
                  path: { type: string }
                  message: { type: string }
...
    Counts:
      type: object
      required: [positions, platformLinks, questions, readings, resumes, logos, departments, teams, locations]
      properties:
        positions: { type: integer, minimum: 0 }
        platformLinks: { type: integer, minimum: 0 }
        questions: { type: integer, minimum: 0 }
        readings: { type: integer, minimum: 0 }
        resumes: { type: integer, minimum: 0 }
        logos: { type: integer, minimum: 0 }
        departments: { type: integer, minimum: 0 }
        teams: { type: integer, minimum: 0 }
        locations: { type: integer, minimum: 0 }
    SynchronizationPreview:
      type: object
      required: [attemptId, direction, source, destination, conflict, willReplace, willMerge, backupRequired, expiresAt]
      properties:
        attemptId: { type: string, format: uuid }
        direction: { type: string, enum: [offline_to_online, working_copy_to_online] }
        source: { $ref: '#/components/schemas/WorkspaceSide' }
        destination: { $ref: '#/components/schemas/WorkspaceSide' }
        conflict: { type: boolean }
        willReplace: { const: online }
        willMerge: { const: false }
        backupRequired: { const: true }
        notices: { type: array, items: { type: string } }
        expiresAt: { type: string, format: date-time }
    WorkspaceSide:
      type: object
      required: [updatedAt, counts]
      properties:
        revision: { type: [integer, 'null'], minimum: 0 }
        updatedAt: { type: string, format: date-time }
        checksum: { type: [string, 'null'] }
        counts: { $ref: '#/components/schemas/Counts' }
    SynchronizationResult:
      type: object
      required: [attemptId, status, revision, backup, counts, completedAt]
      properties:
        attemptId: { type: string, format: uuid }
        status: { const: completed }
        revision: { type: integer, minimum: 1 }
        backup:
          type: object
          required: [id, createdAt]
          properties:
            id: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
        counts: { $ref: '#/components/schemas/Counts' }
        completedAt: { type: string, format: date-time }